# API_documents/t8hami022-workhoursAPI-1.0.2-swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/t8hami022/workhoursAPI/1.0.2"
  } ],
  "info" : {
    "description" : "This is a simple API",
    "version" : "1.0.2",
    "title" : "Simple Inventory API",
    "contact" : {
      "email" : "you@your-company.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/refresh_token" : {
      "post" : {
        "tags" : [ "developers" ],
        "summary" : "refresh accestokens",
        "operationId" : "refresh",
        "description" : "get new accesstoken and refreshtoken with refreshtoken send in cookies",
        "security" : [ {
          "cookieAuth" : [ ]
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "userID" : {
                      "type" : "string",
                      "description" : "User id"
                    },
                    "token" : {
                      "type" : "string",
                      "example" : "1mx6L2L7AMdEsyKy5LW9s8gt6mBxdICwosVn5sjhbwykOoQJFUru6752dwsj2THN",
                      "description" : "jwt Token"
                    }
                  }
                }
              }
            },
            "headers" : {
              "Set-Cookie" : {
                "schema" : {
                  "type" : "string",
                  "example" : "refreshToken=abcde12345; Path=/; HttpOnly"
                }
              }
            }
          },
          "401" : {
            "description" : "Access token is invalid"
          },
          "403" : {
            "description" : "Access token is missing"
          }
        }
      }
    }
  },
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "cookieAuth" : {
        "type" : "apiKey",
        "in" : "cookie",
        "name" : "refreshToken"
      }
    },
...
  }
}